code coverage script fixes

+++ Deface.NET/tests/Deface.NET.UnitTests/code-coverage-report.ps1
@@ -1,5 +1,18 @@
-Remove-Item -Path ".\TestResults" -Recurse -Force
-Remove-Item -Path ".\coverageReport" -Recurse -Force
+dotnet tool install -g dotnet-reportgenerator-globaltool
+
+if (Test-Path -Path ".\TestResults") {
+    Remove-Item -Path ".\TestResults" -Recurse -Force
+}
+
+if (Test-Path -Path ".\coverageReport") {
+    Remove-Item -Path ".\coverageReport" -Recurse -Force
+}
+
 dotnet test --collect:"XPlat Code Coverage"
-reportgenerator -reports:TestResults/**/coverage.cobertura.xml -targetdir:coverageReport
-Start-Process ./coverageReport/index.html
+& "$env:USERPROFILE\.dotnet\tools\reportgenerator" -reports:TestResults/**/coverage.cobertura.xml -targetdir:./coverageReport -reporttypes:Html
+
+if (Test-Path -Path "./coverageReport/index.html") {
+    Start-Process "./coverageReport/index.html"
+} else {
+    Write-Host "Coverage report not found. Please check the reportgenerator command."
+}
